protected every route using withPageAuthRequired and withApiAuthRequired
