+++ app/(dashboard)/admin/page.tsx
@@ -2,15 +2,18 @@ import Admin from "@/components/Admin";
 import { getUserRole } from "@/utils/actions/user_actions";
 import { getSession } from "@auth0/nextjs-auth0";
 import { redirect } from "next/navigation";
+import { withPageAuthRequired } from "@auth0/nextjs-auth0";
 
-const page = async () => {
-  const session = await getSession();
-  const role = await getUserRole(session?.user.sub);
+export default withPageAuthRequired(
+  async () => {
+    const session = await getSession();
+    const role = await getUserRole(session?.user.sub);
 
-  if (role !== "admin") {
-    return redirect("/");
-  }
+    if (role !== "admin") {
+      return redirect("/");
+    }
 
-  return <Admin />;
-};
-export default page;
+    return <Admin />;
+  },
+  { returnTo: "/admin" }
+);
