+++ app/(dashboard)/blog/add-new-post/page.tsx
@@ -1,10 +1,12 @@
 import AddNewPost from "@/components/AddNewPost";
 import { getSystemPreferences } from "@/utils/actions";
+import { withPageAuthRequired } from "@auth0/nextjs-auth0";
 
-const page = async () => {
-  const { language }: { language: string } = await getSystemPreferences();
+export default withPageAuthRequired(
+  async () => {
+    const { language }: { language: string } = await getSystemPreferences();
 
-  return <AddNewPost language={language} />;
-};
-
-export default page;
+    return <AddNewPost language={language} />;
+  },
+  { returnTo: "/blog/add-new-post" }
+);
