+++ app/(dashboard)/canceled/page.tsx
@@ -1,8 +1,13 @@
-export default function Canceled() {
-  return (
-    <div>
-      <h1>Payment Canceled</h1>
-      <p>Your payment was canceled. Please try again.</p>
-    </div>
-  );
-}
+import { withPageAuthRequired } from "@auth0/nextjs-auth0";
+
+export default withPageAuthRequired(
+  async () => {
+    return (
+      <div>
+        <h1>Payment Canceled</h1>
+        <p>Your payment was canceled. Please try again.</p>
+      </div>
+    );
+  },
+  { returnTo: "/canceled" }
+);
