+++ app/(dashboard)/cart/page.tsx
@@ -1,16 +1,15 @@
 import Cart from "@/components/Cart";
 import { getCart } from "@/utils/actions/cart_actions";
 import { getSession } from "@auth0/nextjs-auth0";
+import { withPageAuthRequired } from "@auth0/nextjs-auth0";
 
-const page = async () => {
-  const session = await getSession();
+export default withPageAuthRequired(
+  async () => {
+    const session = await getSession();
 
-  const data = await getCart(session?.user.sub);
+    const data = await getCart(session?.user.sub);
 
-  return (
-    <>
-      <Cart data={data} />
-    </>
-  );
-};
-export default page;
+    return <Cart data={data} />;
+  },
+  { returnTo: "/cart" }
+);
