--- app/(dashboard)/gallery/page.tsx
@@ -1,4 +1,0 @@
-const page = () => {
-  return <div>page</div>;
-};
-export default page;
