+++ app/(dashboard)/success/page.tsx
@@ -1,15 +1,9 @@
-"use client";
+import Success from "@/components/Success";
+import { withPageAuthRequired } from "@auth0/nextjs-auth0";
 
-import { useSearchParams } from "next/navigation";
-
-export default function Success() {
-  const searchParams = useSearchParams();
-  const session_id = searchParams.get("session_id");
-
-  return (
-    <div>
-      <h1>Payment Successful</h1>
-      <p>Your session ID is: {session_id}</p>
-    </div>
-  );
-}
+export default withPageAuthRequired(
+  async () => {
+    return <Success />;
+  },
+  { returnTo: "/success" }
+);
