--- app/api/old-auth/logout/route.ts
@@ -1,7 +1,0 @@
-import { NextRequest } from "next/server";
-
-export const GET = async (req: NextRequest) => {
-  console.log(req.nextUrl.origin);
-
-  return new Response("Logged out successfully!", { status: 200 });
-};
