+++ app/api/refund/route.ts
@@ -1,3 +1,4 @@
+import { withApiAuthRequired } from "@auth0/nextjs-auth0";
 import { sql } from "@vercel/postgres";
 import { NextRequest, NextResponse } from "next/server";
 import Stripe from "stripe";
@@ -6,7 +7,7 @@ const stripe = new Stripe(process.env.STRIPE_SECRET_KEY!, {
   apiVersion: "2024-04-10", // Use the latest API version
 });
 
-export async function POST(req: NextRequest) {
+export const POST = withApiAuthRequired(async (req: NextRequest) => {
   const { order_id } = await req.json();
 
   const dbRes = await sql`SELECT * FROM orders WHERE order_id=${order_id}`;
@@ -19,4 +20,4 @@ export async function POST(req: NextRequest) {
   console.log("refunded");
 
   return NextResponse.json({ refund }, { status: 200 });
-}
+});
