--- app/login/loading.tsx
@@ -1,4 +1,0 @@
-const loading = () => {
-  return <div>loading...</div>;
-};
-export default loading;
