+++ components/Home.tsx
@@ -1,0 +1,4 @@
+const Home = () => {
+  return <div>Home</div>;
+};
+export default Home;
