+++ components/Nav.tsx
@@ -1,13 +1,12 @@
-import { NavLink } from "./NavLink";
 import LanguageSelector from "./LanguageSelector";
 import ChangeTheme from "./ChangeTheme";
 import { getCartTotalCookie } from "@/utils/actions/cart_actions";
 import CartBtn from "./CartBtn";
 import { getSession } from "@auth0/nextjs-auth0";
 import User from "./User";
+import Link from "next/link";
 
 const Nav = async ({
-  nav,
   lang,
   systemPreferences,
 }: {
@@ -20,9 +19,21 @@ const Nav = async ({
   return (
     <nav className="navigation">
       <div className="navigation-left">
-        {nav.map((link) => {
-          return <NavLink key={link.text} text={link.text} href={link.href} />;
-        })}
+        <Link className="nav-link" href={"/"}>
+          Home
+        </Link>
+        <Link className="nav-link" href={"/store"}>
+          Store
+        </Link>
+        <Link className="nav-link" href={"/blog"}>
+          Blog
+        </Link>
+        <Link className="nav-link" href={"/premium"}>
+          Premium
+        </Link>
+        <Link className="nav-link" href={"/contact"}>
+          Contact
+        </Link>
       </div>
       <div className="navigation-right">
         <CartBtn cart_total={total} />
