+++ components/UserDropdown.tsx
@@ -8,6 +8,7 @@ const UserDropdown = () => {
       <Link href="/my-products">my products</Link>
       <Link href="/my-posts">my posts</Link>
       <Link href="/orders">orders</Link>
+      <Link href="/admin">admin</Link>
       <a href="/api/auth/logout">Logout</a>
     </div>
   );
